+++ CHANGELOG.md
@@ -8,8 +8,8 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 - `hasResourceInfo`: a function that can verify whether its parameter (e.g. a file or a
   SolidDataset) was fetched from somewhere, or was initialised in-memory.
-- `createEmptyContainerAt` and `createEmptyContainerInContainer`: two functions that can help you
-  create an empty Container at a given location or in another Container on the Pod, respectively.
+- `createContainerAt` and `createContainerInContainer`: two functions that can help you create an
+  empty Container at a given location or in another Container on the Pod, respectively.
 - `isThing`: a function that can verify whether its parameter is a Thing.
 - `mockSolidDatasetFrom`, `mockContainerFrom`, `mockFileFrom`, `mockThingFrom`,
   `addMockResourceAclTo` and `addMockFallbackAclTo`: functions that allow you to mock the